% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_read_csv.R
\name{simple_read_csv}
\alias{simple_read_csv}
\title{Simple Read CSV
This function is meant to create a CSV/TSV reading function that makes more sense than the default read.csv and read.csv2 functions}
\usage{
simple_read_csv(separator, decimal, theFile)
}
\arguments{
\item{separator}{This indicates which character is being used to demlimit a CSV. Accepted values are comma, semicolon, and tab.}

\item{decimal}{This indicates which character is being used to indicate decimal places in a number. Acceptable values are a comma or period/full stop.}

\item{theFile}{The name of the CSV or TSV.}
}
\description{
Simple Read CSV
This function is meant to create a CSV/TSV reading function that makes more sense than the default read.csv and read.csv2 functions
}
\examples{
simple_read_csv(separator = '\\t', decimal = ',', theFile = 'bob.csv')
}
\keyword{csv}
\keyword{read}
\keyword{simple}
\keyword{tsv}
